# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LendingOS — Multi-Tenant AI Lending Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers/{applicationId}:
    get:
      operationId: getOffers
      tags: [offers]
      summary: Get loan offers for application
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Loan offers from multiple NBFCs/Banks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanOffer"
  /offers/{applicationId}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Accept a loan offer
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferAcceptanceInput"
      responses:
        "200":
          description: Offer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /offers/calculate-emi:
    post:
      operationId: calculateEmi
      tags: [offers]
      summary: Calculate EMI for loan parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiCalculationInput"
      responses:
        "200":
          description: EMI calculation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiCalculation"

  # ─── LOANS ──────────────────────────────────────────────────────────────────
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List active loans
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, closed, npa, written_off]
        - name: customerId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"
  /loans/{loanId}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan details
      parameters:
        - name: loanId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/schedule:
    get:
      operationId: getLoanSchedule
      tags: [loans]
      summary: Get loan repayment schedule
      parameters:
        - name: loanId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Repayment schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepaymentScheduleItem"

  # ─── REPAYMENTS ─────────────────────────────────────────────────────────────
  /repayments:
    get:
      operationId: listRepayments
      tags: [repayments]
      summary: List repayments
      parameters:
        - name: loanId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue, partial]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Repayments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepaymentListResponse"
  /repayments/{repaymentId}/record:
    post:
      operationId: recordRepayment
      tags: [repayments]
      summary: Record a repayment
      parameters:
        - name: repaymentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentRecordInput"
      responses:
        "200":
          description: Repayment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repayment"

  # ─── COLLECTIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoanApplication:
      type: object
      required: [id, customerId, tenantId, productId, requestedAmount, requestedTenure, purpose, status, createdAt]
      properties:
        id: { type: string }
        applicationNumber: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        tenantId: { type: string }
        tenantName: { type: string }
        productId: { type: string }
        productName: { type: string }
        requestedAmount: { type: number }
        requestedTenure: { type: integer }
        purpose: { type: string }
        status:
          type: string
          enum: [draft, submitted, under_review, kyc_pending, kyc_verified, risk_assessment, offer_generated, offer_accepted, esign_pending, approved, disbursed, rejected, withdrawn]
        riskScore: { type: ["number", "null"] }
        riskGrade: { type: ["string", "null"] }
        approvedAmount: { type: ["number", "null"] }
        approvedTenure: { type: ["integer", "null"] }
        approvedRate: { type: ["number", "null"] }
        rejectionReason: { type: ["string", "null"] }
        disbursedAt: { type: ["string", "null"] }
        bankAccount: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    LoanOffer:
      type: object
      required: [id, applicationId, tenantId, tenantName, offeredAmount, tenure, interestRate, emi, processingFee, expiresAt]
      properties:
        id: { type: string }
        applicationId: { type: string }
        tenantId: { type: string }
        tenantName: { type: string }
        tenantLogo: { type: ["string", "null"] }
        offeredAmount: { type: number }
        tenure: { type: integer }
        interestRate: { type: number }
        emi: { type: number }
        processingFee: { type: number }
        totalInterest: { type: number }
        totalRepayable: { type: number }
        approvalProbability: { type: number }
        disbursementTime: { type: string }
        expiresAt: { type: string }
        isAccepted: { type: boolean }
    OfferAcceptanceInput:
      type: object
      required: [offerId]
      properties:
        offerId: { type: string }
    EmiCalculationInput:
      type: object
      required: [principal, annualRate, tenureMonths]
      properties:
        principal: { type: number }
        annualRate: { type: number }
        tenureMonths: { type: integer }
    EmiCalculation:
      type: object
      required: [emi, totalPayable, totalInterest, processingFee]
      properties:
        emi: { type: number }
        totalPayable: { type: number }
        totalInterest: { type: number }
        processingFee: { type: number }
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/RepaymentScheduleItem"

    # ── LOANS ─────────────────────────────────────────────────────────────────
    Loan:
      type: object
      required: [id, applicationId, customerId, tenantId, principalAmount, outstandingAmount, interestRate, tenure, emiAmount, status, disbursedAt]
      properties:
        id: { type: string }
        loanNumber: { type: string }
        applicationId: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        tenantId: { type: string }
        principalAmount: { type: number }
        outstandingAmount: { type: number }
        totalPaid: { type: number }
        interestRate: { type: number }
        tenure: { type: integer }
        emiAmount: { type: number }
        nextEmiDate: { type: ["string", "null"] }
        status:
          type: string
          enum: [active, closed, npa, written_off]
        dpd: { type: integer }
        disbursedAt: { type: string }
        closedAt: { type: ["string", "null"] }
    LoanListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    RepaymentScheduleItem:
      type: object
      required: [installmentNumber, dueDate, emiAmount, principal, interest, outstandingAfter]
      properties:
        installmentNumber: { type: integer }
        dueDate: { type: string }
        emiAmount: { type: number }
        principal: { type: number }
        interest: { type: number }
        outstandingAfter: { type: number }
        status:
          type: string
          enum: [pending, paid, overdue, partial]
        paidAt: { type: ["string", "null"] }
        paidAmount: { type: ["number", "null"] }

    # ── REPAYMENTS ────────────────────────────────────────────────────────────
    Repayment:
      type: object
      required: [id, loanId, dueDate, emiAmount, status]
      properties:
        id: { type: string }
        loanId: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        installmentNumber: { type: integer }
        dueDate: { type: string }
        emiAmount: { type: number }
        paidAmount: { type: ["number", "null"] }
        paidAt: { type: ["string", "null"] }
        paymentMode: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, paid, overdue, partial]
        dpd: { type: integer }
        penaltyAmount: { type: number }
    RepaymentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Repayment"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    RepaymentRecordInput:
      type: object
      required: [paidAmount, paymentMode]
      properties:
        paidAmount: { type: number }
        paymentMode:
          type: string
          enum: [cash, upi, neft, imps, cheque, auto_debit]
        transactionId: { type: string }
        paidAt: { type: string }

    # ── COLLECTIONS ───────────────────────────────────────────────────────────
...